services/provider-adapters/mock-llm/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
prometheus-client==0.21.1
opentelemetry-api==1.32.1
opentelemetry-sdk==1.32.1
opentelemetry-exporter-otlp-proto-grpc==1.32.1
opentelemetry-instrumentation-fastapi==0.53b1
